Pushing working version before deleting all the extra work
