+++ chals/xorish/ctfsh.yml
@@ -1,10 +1,8 @@
 challenge:
-  name: "XORish"
+  name: "xorish"
   author: "quasar"
   category: "Crypto"
   description: "I got a scrambled message"
-  flag:
-    value: "flag{x0r_x0r}"
-  download:
-    files:
-      - xorish.pcap
+  flag: "flag{x0r_x0r}"
+  downloads:
+    - xorish.pcap
