+++ deploy8/chal-old/Dockerfile
@@ -1,0 +1,5 @@
+FROM python:3.11-slim
+WORKDIR /app
+COPY app.py .
+ENV FLAG=CTF{default}
+CMD ["python", "app.py"]
