+++ deploy8/chal/ctfsh.yml
@@ -1,0 +1,10 @@
+challenge:
+  name: "Elementary"
+  author: "B00TK1D"
+  category: "Crypto"
+  description: "I made a calculator for elementary students.  I made sure it can only do basic operations, so it should be safe."
+  flag:
+    value: "cube{3l3m3nt4ry_mY_d34r_w47s0n_181c2f60}"
+  download:
+    files:
+      - main.py
