+++ deploy8/v3/util.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"log"
+	"runtime/debug"
+)
+
+func must(err error) {
+	if err != nil {
+		debug.PrintStack()
+		log.Fatal(err)
+	}
+}
